+++ packages/cli/templates/webcomponents/igc-ts/accordion/default/index.ts
@@ -1,0 +1,19 @@
+import { TypeScriptFileUpdate } from "@igniteui/cli-core";
+import { IgniteUIForWebComponentsTemplate } from "../../../../../lib/templates/IgniteUIForWebComponentsTemplate";
+
+class IgcAccordionTemplate extends IgniteUIForWebComponentsTemplate {
+	constructor() {
+		super(__dirname);
+		this.components = ["Accordion"];
+		this.controlGroup = "Layouts";
+		this.listInComponentTemplates = true;
+		this.id = "accordion";
+		this.projectType = "igc-ts";
+		this.name = "Accordion";
+		this.description = "basic IgcAccordion";
+	}
+	protected addClassDeclaration(mainModule: TypeScriptFileUpdate, projPath: string, name: string, modulePath: string) {
+		// not applicable with custom module
+	}
+}
+module.exports = new IgcAccordionTemplate();