+++ packages/cli/templates/webcomponents/igc-ts/projects/_base/files/src/app/app-routing.ts
@@ -1,11 +1,7 @@
-import {
-  Route
-} from '@vaadin/router';
-import './home/home.js';
+import { Route } from '@vaadin/router';
 import './not-found/not-found.js';
 
 export const routes: Route[] = [
-  { path: '/', component: 'app-home', name: 'Home' },
   // The fallback route should always be after other alternatives.
   { path: '(.*)', component: 'app-not-found' },
 ];